restructure project and update tests
