+++ pyproject.toml
@@ -36,6 +36,6 @@ line-length = 100
 
 [tool.ruff.format]
 quote-style = "double"
-indent-style = "tab"
+indent-style = "space"
 docstring-code-format = true
 docstring-code-line-length = 80
