+++ src/backend/routes/temp.py
@@ -1,0 +1,15 @@
+"""
+Temporary route for initial setup
+"""
+
+from aws_lambda_powertools.event_handler.api_gateway import Router
+
+router = Router()
+
+
+@router.get("/test")
+def thingy():
+    """
+    Dummy route for use in initial project structure setup. Can be removed later.
+    """
+    return {"Hello": "World"}
