--- test/backend/unit/test_handler.py
@@ -1,4 +1,0 @@
-from backend.handler import lambda_handler
-	
-def test_endpoint(get_request):
-	print(lambda_handler(event=get_request[0], context=get_request[1]))
